+++ Datatexts/CombatIndicator.lua
@@ -1,10 +1,11 @@
-local E, L, V, P, G = unpack(ElvUI); --Inport: Engine, Locales, PrivateDB, ProfileDB, GlobalDB
+local E = unpack(ElvUI); --Inport: Engine, Locales, PrivateDB, ProfileDB, GlobalDB
 local DT = E:GetModule('DataTexts')
+local EDT = select(2, ...).EDT
+
 local format = format
 
-local function OnEvent(self, event, ...)
+local function OnEvent(self, event)
 
-	
 	if event == 'PLAYER_REGEN_ENABLED' then
 		self.text:SetText('Out of Combat')
 		return;
@@ -12,15 +13,15 @@ local function OnEvent(self, event, ...)
 		self.text:SetText(format('|cffff0000In Combat|r'))
 		return;
 	end
-	
+
 	self.text:SetText('Out of Combat')
 end
-	
+
 --[[
 	DT:RegisterDatatext(name, events, eventFunc, updateFunc, clickFunc, onEnterFunc, onLeaveFunc)
-	
+
 	name - name of the datatext (required)
-	events - must be a table with string values of event names to register 
+	events - must be a table with string values of event names to register
 	eventFunc - function that gets fired when an event gets triggered
 	updateFunc - onUpdate script target function
 	click - function to fire when clicking the datatext
@@ -29,11 +30,9 @@ end
 ]]
 
 local events = {
-	'PLAYER_ENTERING_WORLD',
-	'ELVUI_FORCE_UPDATE',
 	'PLAYER_REGEN_ENABLED',
 	'PLAYER_REGEN_DISABLED'
 }
 
-DT:RegisterDatatext('Combat Indicator', 'ExtraDataTexts', events, OnEvent)
+DT:RegisterDatatext('Combat Indicator', EDT.DTCategory, events, OnEvent)